+++ js/jquery.terminal.js
@@ -31,7 +31,7 @@
  * Copyright (c) 2007-2013 Alexandru Marasteanu <hello at alexei dot ro>
  * licensed under 3 clause BSD license
  *
- * Date: Sat, 09 Sep 2017 15:24:45 +0000
+ * Date: Sun, 10 Sep 2017 07:58:01 +0000
  */
 
 /* TODO:
@@ -1339,7 +1339,7 @@
         function paste_event() {
             clip.val('');
             paste_count = 0;
-            clip.focus();
+            clip.trigger('focus', [true]);
             clip.on('input', function input(e) {
                 paste(e);
                 clip.off('input', input);
@@ -1775,7 +1775,7 @@
             if (self.isenabled()) {
                 var clip = self.find('textarea');
                 if (!clip.is(':focus')) {
-                    clip.focus();
+                    clip.trigger('focus', [true]);
                 }
                 //wait until Browser insert text to textarea
                 self.oneTime(100, function() {
@@ -6284,11 +6284,13 @@
                                                  settings.login);
         }
         terminals.append(self);
-        //terminals.set(self);
         self.on('focus.terminal', 'textarea', function(e, skip) {
-            if (!enabled && !skip) {
-                self.enable();
-            }
+            // for cases when user press tab to focus terminal
+            self.oneTime(1, function() {
+                if (!self.enable() && !skip) {
+                    self.enable();
+                }
+            });
         });
         function focus_terminal() {
             if (old_enabled) {